% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/met2model.LINKAGES.R
\name{met2model.LINKAGES}
\alias{met2model.LINKAGES}
\title{Write LINKAGES met files}
\usage{
met2model.LINKAGES(in.path, in.prefix, outfolder, start_date, end_date, ...,
  overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{in.path}{path on disk where CF file lives}

\item{in.prefix}{prefix for each file}

\item{outfolder}{location where model specific output is written.}
}
\value{
OK if everything was succesful.
}
\description{
Converts a met CF file to a model specific met file. The input
files are calld <in.path>/<in.prefix>.YYYY.cf
}
\author{
Ann Raiho, Betsy Cowdery
}

