% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/model2netcdf.LINKAGES.R
\name{model2netcdf.MODEL}
\alias{model2netcdf.LINKAGES}
\alias{model2netcdf.MODEL}
\title{Code to convert MODELS's output into netCDF format}
\usage{
model2netcdf.LINKAGES(outdir, sitelat, sitelon, start_date = NULL,
  end_date = NULL, force = FALSE)
}
\arguments{
\item{outdir}{Location of model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}
}
\description{
Convert MODEL output into the NACP Intercomparison format (ALMA using netCDF)
}
\author{
Ann Raiho
}

