% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.restart.LINKAGES.R
\name{read.restart.LINKAGES}
\alias{read.restart.LINKAGES}
\title{read.restart.LINKAGES}
\usage{
read.restart.LINKAGES(outdir, runid, stop.time, settings, var.names = NULL,
  params = NULL)
}
\arguments{
\item{outdir}{output directory}

\item{runid}{run ID}

\item{stop.time}{year that is being read}

\item{var.names}{var.names to be extracted}

\item{multi.settings}{PEcAn settings object}
}
\value{
X.vec      vector of forecasts
}
\description{
Read Restart for LINKAGES
}
\author{
Ann Raiho \email{araiho@nd.edu}
}

