% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.inputs.LINKAGES.R
\name{split.inputs.LINKAGES}
\alias{split.inputs.LINKAGES}
\title{split.inputs.LINKAGES}
\usage{
\method{split}{inputs.LINKAGES}(settings, start.time, stop.time)
}
\value{
files split up climate files
}
\description{
Splits climate met for LINKAGES
}
\author{
Ann Raiho
}

