% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_inputs.LINKAGES.R
\name{split_inputs.LINKAGES}
\alias{split_inputs.LINKAGES}
\title{split_inputs.LINKAGES}
\usage{
split_inputs.LINKAGES(settings, start.time, stop.time, inputs)
}
\value{
files split up climate files
}
\description{
Splits climate met for LINKAGES
}
\author{
Ann Raiho
}
