\name{write.restart.LINKAGES}
\alias{write.restart.LINKAGES}
\title{write.restart.LINKAGES}
\usage{
write.restart.LINKAGES(out.dir, runid, time, settings, analysis.vec,
  RENAME = TRUE, PLOT = FALSE, variables, sample_parameters = FALSE,
  trait.values = NA)
}
\arguments{
  \item{out.dir}{output directory}

  \item{runid}{run ID}

  \item{time}{year that is being read}

  \item{settings}{PEcAn settings object}

  \item{analysis.vec}{analysis vector}

  \item{RENAME}{flag to either rename output file or not}

  \item{PLOT}{flag to make plots or not}

  \item{variables}{}

  \item{sample_parameters}{}

  \item{trait.values}{}
}
\value{
NONE
}
\description{
Write restart files for LINKAGES
}
\author{
Ann Raiho \email{araiho@nd.edu}
}

