\name{read.restart.LINKAGES}
\alias{read.restart.LINKAGES}
\title{read.restart.LINKAGES}
\usage{
read.restart.LINKAGES(outdir, runid, time, settings, variables)
}
\arguments{
  \item{outdir}{output directory}

  \item{runid}{run ID}

  \item{time}{year that is being read}

  \item{settings}{PEcAn settings object}

  \item{variables}{variables to be extracted}
}
\value{
X.vec vector of forecasts
}
\description{
Read Restart for LINKAGES
}
\author{
Ann Raiho \email{araiho@nd.edu}
}

