% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.restart.LINKAGES}
\alias{read.restart.LINKAGES}
\title{read.restart.LINKAGES}
\usage{
read.restart.LINKAGES(outdir, runid, stop.time, multi.settings, var.names,
  params = NULL)
}
\arguments{
\item{outdir}{output directory}

\item{runid}{run ID}

\item{stop.time}{year that is being read}

\item{multi.settings}{PEcAn settings object}

\item{var.names}{var.names to be extracted}
}
\value{
X.vec      vector of forecasts
}
\description{
Read Restart for LINKAGES
}
\author{
Ann Raiho \email{araiho@nd.edu}
}

