% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_restart.LINKAGES.R
\name{read_restart.LINKAGES}
\alias{read_restart.LINKAGES}
\title{Read Restart for LINKAGES}
\usage{
read_restart.LINKAGES(
  outdir,
  runid,
  stop.time,
  settings,
  var.names = NULL,
  params = NULL
)
}
\arguments{
\item{outdir}{output directory}

\item{runid}{run ID}

\item{stop.time}{year that is being read}

\item{settings}{PEcAn settings object}

\item{var.names}{var.names to be extracted}

\item{params}{passed on to return value}
}
\value{
X.vec      vector of forecasts
}
\description{
Read Restart for LINKAGES
}
\author{
Ann Raiho \email{araiho@nd.edu}
}
