% Generated by roxygen2 (4.0.2): do not edit by hand
\name{split.inputs.LINKAGES}
\alias{split.inputs.LINKAGES}
\title{split.inputs.LINKAGES}
\usage{
split.inputs.LINKAGES(inputs, start.time, stop.time)
}
\arguments{
\item{inputs}{settings$run$inputs}

\item{start.time}{}

\item{stop.time}{}
}
\value{
files split up climate files
}
\description{
Splits climate met for LINKAGES
}
\author{
Mike Dietze
}

