% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.LINKAGES.R
\name{write.config.LINKAGES}
\alias{write.config.LINKAGES}
\title{Writes a LINKAGES config file.}
\usage{
write.config.LINKAGES(
  defaults = NULL,
  trait.values,
  settings,
  run.id,
  restart = NULL,
  spinup = FALSE,
  inputs = NULL,
  IC = NULL
)
}
\arguments{
\item{defaults}{list of defaults to process}

\item{trait.values}{vector of samples for a given trait}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}

\item{restart}{logical: Write a restart file?
If NULL (default), treated as FALSE}

\item{spinup}{logical: perform spinup using `spinup.LINKAGES()`?
If NULL (default), treated as FALSE}

\item{inputs}{inputs section of a PEcAn settings object.
Currently only used for climate file (inputs$met$path),
which is taken from `settings$input$met$path` if `inputs` is NULL.}

\item{IC}{TODO currently ignored}
}
\value{
configuration file for LINKAGES for given run
}
\description{
Requires a pft xml object, a list of trait values for a single model run,
and the name of the file to create
}
\author{
Ann Raiho, Betsy Cowdery
}
