% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.restart.LINKAGES}
\alias{write.restart.LINKAGES}
\title{write.restart.LINKAGES}
\usage{
write.restart.LINKAGES(outdir, runid, time, settings, analysis.vec, variables,
  sample_parameters = FALSE, trait.values = NA, met = NULL,
  RENAME = TRUE)
}
\arguments{
\item{outdir}{output directory}

\item{runid}{run ID}

\item{time}{year that is being read}

\item{settings}{PEcAn settings object}

\item{analysis.vec}{analysis vector}

\item{variables}{}

\item{sample_parameters}{}

\item{trait.values}{}

\item{RENAME}{flag to either rename output file or not}
}
\value{
NONE
}
\description{
Write restart files for LINKAGES
}
\author{
Ann Raiho \email{araiho@nd.edu}
}

